A_service/bus-destination_service/requirements.txt
Flask==2.2.5
flask_cors==4.0.0
mysql-connector-python==8.4.0